# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DJD Agent Score API",
    "description": "Reputation scoring for autonomous AI agent wallets on Base. Scores are derived from on-chain USDC behavioral data and optional self-registration. Paid endpoints require an x402 USDC micropayment on Base; free endpoints require no authentication. Interactive docs available at /docs.",
    "version": "2.0.0",
    "contact": { "email": "feedback@djdagentscore.dev" },
    "license": { "name": "MIT", "url": "https://github.com/jacobsd32-cpu/djdagentscore/blob/main/LICENSE" }
  },
  "servers": [
    { "url": "https://djdagentscore.dev", "description": "Production" }
  ],
...
  "paths": {
...
    "/v1/monitor": {
      "post": {
        "tags": ["monitoring"],
        "summary": "Create monitoring subscription",
        "description": "Creates a managed monitoring subscription for a target wallet and provisions the underlying webhook delivery automatically. Policies can target any wallet; delivery is sent to your HTTPS endpoint. `score_monitoring` supports `threshold_score`. Forensics policies support `forensics_filter`. `anomaly_monitoring` delivers score-drop, score-spike, balance-freefall, and sybil-flag alerts. **Requires Bearer API key authentication.**",
        "operationId": "createMonitoringSubscription",
        "security": [{ "bearerApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "policy_type"],
                "properties": {
                  "target_wallet": {
                    "type": "string",
                    "pattern": "^0x[0-9a-fA-F]{40}$",
                    "description": "Wallet to monitor. Defaults to the authenticated wallet when omitted."
                  },
                  "policy_type": {
                    "type": "string",
                    "enum": ["score_monitoring", "forensics_monitoring", "forensics_disputes", "forensics_watchlist", "anomaly_monitoring"],
                    "description": "Managed monitoring policy to provision."
                  },
                  "url": {
                    "type": "string",
                    "format": "uri",
                    "description": "HTTPS endpoint to receive monitoring webhook deliveries."
                  },
                  "threshold_score": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Optional score threshold for `score_monitoring`. Defaults to 60 when omitted."
                  },
                  "forensics_filter": {
                    "$ref": "#/components/schemas/WebhookForensicsFilter"
                  }
                }
              },
              "example": {
                "target_wallet": "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528",
                "policy_type": "forensics_monitoring",
                "url": "https://example.com/monitoring/djd",
                "forensics_filter": {
                  "minimum_risk_level": "elevated",
                  "reasons": ["payment_fraud"]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Monitoring subscription created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MonitoringSubscriptionCreatedResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": {
            "description": "Maximum monitoring subscriptions per subscriber exceeded",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      },
      "get": {
        "tags": ["monitoring"],
        "summary": "List monitoring subscriptions",
        "description": "Lists all managed monitoring subscriptions owned by the authenticated API-key subscriber. **Requires Bearer API key authentication.**",
        "operationId": "listMonitoringSubscriptions",
        "security": [{ "bearerApiKey": [] }],
        "responses": {
          "200": {
            "description": "Monitoring subscription list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscriptions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MonitoringSubscription" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/monitor/{id}": {
      "delete": {
        "tags": ["monitoring"],
        "summary": "Delete monitoring subscription",
        "description": "Deactivates a managed monitoring subscription and its underlying webhook delivery configuration. **Requires Bearer API key authentication.**",
        "operationId": "deleteMonitoringSubscription",
        "security": [{ "bearerApiKey": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Monitoring subscription ID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Monitoring subscription deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Monitoring subscription not found or already disabled",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/v1/forensics/summary": {
      "get": {
        "tags": ["forensics"],
        "summary": "DJD Forensics overview",
        "description": "Returns a wallet-level forensics summary: active fraud report counts, total penalties, reporter breadth, dispute status, and the most recent incidents. **Requires x402 payment of $0.10 USDC on Base or Bearer API key.**",
        "operationId": "getForensicsOverview",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.10", "network": "base" },
        "parameters": [
          { "$ref": "#/components/parameters/wallet" }
        ],
        "responses": {
          "200": {
            "description": "Forensics overview returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ForensicsOverviewResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/forensics/dispute": {
      "post": {
        "tags": ["forensics"],
        "summary": "Open a fraud-report dispute",
        "description": "Allows the reported wallet to open a dispute against a specific fraud report. The disputing wallet is derived from the x402 payment or API key identity and must match the report target wallet. **Requires x402 payment of $0.05 USDC on Base or Bearer API key.**",
        "operationId": "createForensicsDispute",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.05", "network": "base" },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FraudDisputeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fraud dispute opened",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FraudDisputeResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "404": {
            "description": "Fraud report not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "409": {
            "description": "Dispute already exists or wallet is not authorized to dispute this report",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/forensics/feed": {
      "get": {
        "tags": ["forensics"],
        "summary": "DJD Forensics incident feed",
        "description": "Returns recent fraud-report incidents across the corpus. Supports reason and date-range filtering plus configurable result limits. **Requires x402 payment of $0.30 USDC on Base or Bearer API key.**",
        "operationId": "getForensicsFeed",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.30", "network": "base" },
        "parameters": [
          {
            "name": "reason",
            "in": "query",
            "required": false,
            "description": "Optional report reason filter",
            "schema": { "type": "string", "enum": ["failed_delivery", "payment_fraud", "impersonation", "malicious_behavior", "other"] }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of incidents to return (1-100, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Return incidents after this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return incidents before this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Forensics incident feed returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ForensicsFeedResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/forensics/watchlist": {
      "get": {
        "tags": ["forensics"],
        "summary": "DJD Forensics watchlist",
        "description": "Returns a ranked watchlist of the most-reported wallets across the corpus. Supports date-range filtering and configurable result limits. **Requires x402 payment of $0.25 USDC on Base or Bearer API key.**",
        "operationId": "getForensicsWatchlist",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.25", "network": "base" },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of wallets to return (1-100, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Return incidents after this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return incidents before this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Forensics watchlist returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ForensicsWatchlistResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/forensics/reports": {
      "get": {
        "tags": ["forensics"],
        "summary": "DJD Forensics incident feed",
        "description": "Returns fraud-report incidents for a wallet with full report details. Supports date-range filtering and configurable result limits. **Requires x402 payment of $0.15 USDC on Base or Bearer API key.**",
        "operationId": "getForensicsReports",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.15", "network": "base" },
        "parameters": [
          { "$ref": "#/components/parameters/wallet" },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of reports to return (1-100, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Return reports after this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return reports before this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Forensics report feed returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ForensicsReportsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/forensics/timeline": {
      "get": {
        "tags": ["forensics"],
        "summary": "DJD Forensics merged timeline",
        "description": "Returns a merged timeline of score snapshots and fraud incidents for a wallet. Supports date-range filtering and configurable result limits. **Requires x402 payment of $0.20 USDC on Base or Bearer API key.**",
        "operationId": "getForensicsTimeline",
        "security": [{ "x402": [] }, { "bearerApiKey": [] }],
        "x-x402": { "price": "$0.20", "network": "base" },
        "parameters": [
          { "$ref": "#/components/parameters/wallet" },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of combined events to return (1-100, default 50)",
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "description": "Return events after this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "description": "Return events before this ISO 8601 date (YYYY-MM-DD)",
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Merged forensics timeline returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ForensicsTimelineResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "404": {
            "description": "No forensics data found for this wallet",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "wallet": {
        "name": "wallet",
        "in": "query",
        "required": true,
        "description": "Base wallet address (EIP-55 checksummed or lowercase)",
        "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{40}$" },
        "example": "0x3e4ef1f774857c69e33dddc471e110c7ac7bb528"
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "description": "Machine-readable error code" },
              "message": { "type": "string", "description": "Human-readable error message" },
              "details": { "type": "object", "description": "Optional additional context", "additionalProperties": true }
            },
            "required": ["code", "message"]
          }
        }
      },
...
      "FraudDisputeRequest": {
        "type": "object",
        "required": ["report_id", "reason", "details"],
        "properties": {
          "report_id": { "type": "string", "description": "Fraud report ID from the reports or feed endpoints" },
          "reason": {
            "type": "string",
            "enum": ["fulfilled_service", "mistaken_identity", "resolved_offchain", "inaccurate_report", "other"]
          },
          "details": { "type": "string", "description": "Supporting explanation or evidence for the dispute" }
        }
      },
      "FraudDisputeResponse": {
        "type": "object",
        "properties": {
          "disputeId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["open"] },
          "reportId": { "type": "string" },
          "targetWallet": { "type": "string" }
        }
      },
...
      "WebhookForensicsFilter": {
        "type": "object",
        "nullable": true,
        "properties": {
          "minimum_risk_level": {
            "type": "string",
            "enum": ["clear", "watch", "elevated", "critical"],
            "description": "Only emit Forensics events whose resulting risk level meets or exceeds this threshold."
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["failed_delivery", "payment_fraud", "impersonation", "malicious_behavior", "other"]
            },
            "description": "Only emit Forensics events tied to these report reasons."
          }
        }
      },
...
      "MonitoringSubscription": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "target_wallet": { "type": "string" },
          "policy_type": {
            "type": "string",
            "enum": ["score_monitoring", "forensics_monitoring", "forensics_disputes", "forensics_watchlist", "anomaly_monitoring"]
          },
          "url": { "type": "string" },
          "events": { "type": "array", "items": { "type": "string" } },
          "threshold_score": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
          "forensics_filter": { "$ref": "#/components/schemas/WebhookForensicsFilter" },
          "is_active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "disabled_at": { "type": "string", "format": "date-time", "nullable": true },
          "failure_count": { "type": "integer" },
          "last_delivery_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "MonitoringSubscriptionCreatedResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/MonitoringSubscription" },
          {
            "type": "object",
            "properties": {
              "secret": {
                "type": "string",
                "description": "HMAC secret for verifying webhook signatures — returned ONLY on creation"
              },
              "presets_applied": { "type": "array", "items": { "type": "string" } },
              "message": { "type": "string" }
            }
          }
        ]
      },
...
      "ForensicsOverviewResponse": {
        "type": "object",
        "properties": {
          "wallet": { "type": "string" },
          "risk_level": { "type": "string", "enum": ["clear", "watch", "elevated", "critical"] },
          "current_score": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
          "current_tier": { "type": "string", "nullable": true },
          "report_count": { "type": "integer" },
          "total_penalty_applied": { "type": "integer" },
          "unique_reporters": { "type": "integer" },
          "most_recent_report_at": { "type": "string", "format": "date-time", "nullable": true },
          "dispute_status": { "type": "string", "enum": ["none", "open", "resolved"] },
          "open_disputes": { "type": "integer" },
          "resolved_disputes": { "type": "integer" },
          "score_history_entries": { "type": "integer" },
          "reasons": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "reason": { "type": "string" },
                "count": { "type": "integer" }
              }
            }
          },
          "recent_reports": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "report_id": { "type": "string" },
                "reason": { "type": "string" },
                "created_at": { "type": "string", "format": "date-time" },
                "penalty_applied": { "type": "integer" }
              }
            }
          }
        }
      },
...
      "ForensicsReportsResponse": {
        "type": "object",
        "properties": {
          "wallet": { "type": "string" },
          "risk_level": { "type": "string", "enum": ["clear", "watch", "elevated", "critical"] },
          "reports": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "report_id": { "type": "string" },
                "reason": { "type": "string" },
                "details": { "type": "string" },
                "created_at": { "type": "string", "format": "date-time" },
                "penalty_applied": { "type": "integer" }
              }
            }
          },
          "count": { "type": "integer", "description": "Total reports matching the query" },
          "returned": { "type": "integer", "description": "Number of reports in this response" },
          "unique_reporters": { "type": "integer" },
          "total_penalty_applied": { "type": "integer" },
          "period": {
            "type": "object",
            "properties": {
              "from": { "type": "string", "format": "date-time", "nullable": true },
              "to": { "type": "string", "format": "date-time", "nullable": true }
            }
          }
        }
      },
      "ForensicsWatchlistResponse": {
        "type": "object",
        "properties": {
          "wallets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "rank": { "type": "integer" },
                "wallet": { "type": "string" },
                "risk_level": { "type": "string", "enum": ["clear", "watch", "elevated", "critical"] },
                "current_score": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
                "current_tier": { "type": "string", "nullable": true },
                "report_count": { "type": "integer" },
                "unique_reporters": { "type": "integer" },
                "total_penalty_applied": { "type": "integer" },
                "most_recent_report_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "count": { "type": "integer", "description": "Total wallets matching the query" },
          "returned": { "type": "integer", "description": "Number of wallets in this response" },
          "period": {
            "type": "object",
            "properties": {
              "from": { "type": "string", "format": "date-time", "nullable": true },
              "to": { "type": "string", "format": "date-time", "nullable": true }
            }
          }
        }
      },
      "ForensicsFeedResponse": {
        "type": "object",
        "properties": {
          "incidents": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "report_id": { "type": "string" },
                "wallet": { "type": "string" },
                "reason": { "type": "string" },
                "details": { "type": "string" },
                "created_at": { "type": "string", "format": "date-time" },
                "penalty_applied": { "type": "integer" },
                "current_score": { "type": "integer", "minimum": 0, "maximum": 100, "nullable": true },
                "current_tier": { "type": "string", "nullable": true },
                "risk_level": { "type": "string", "enum": ["clear", "watch", "elevated", "critical"] },
                "report_count": { "type": "integer" },
                "unique_reporters": { "type": "integer" },
                "total_penalty_applied": { "type": "integer" }
              }
            }
          },
          "count": { "type": "integer", "description": "Total incidents matching the query" },
          "returned": { "type": "integer", "description": "Number of incidents in this response" },
          "reason_filter": { "type": "string", "nullable": true },
          "period": {
            "type": "object",
            "properties": {
              "from": { "type": "string", "format": "date-time", "nullable": true },
              "to": { "type": "string", "format": "date-time", "nullable": true }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request parameters",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "PaymentRequired": {
        "description": "x402 payment required. The response includes a WWW-Authenticate header with payment details. Use an x402-compatible client to pay and retry.",
        "headers": {
          "WWW-Authenticate": { "schema": { "type": "string" } }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded (120 requests/hour per payer wallet on paid endpoints)",
        "headers": {
          "RateLimit-Limit": { "schema": { "type": "integer" }, "description": "Maximum requests per window" },
          "RateLimit-Remaining": { "schema": { "type": "integer" }, "description": "Requests remaining in current window" },
          "RateLimit-Reset": { "schema": { "type": "string", "format": "date-time" }, "description": "When the current window resets" },
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until rate limit resets" }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "securitySchemes": {
      "x402": {
        "type": "http",
        "scheme": "x402",
        "description": "x402 micropayment protocol (EIP-3009 USDC transferWithAuthorization on Base). See https://www.x402.org for client libraries."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Admin API key for management endpoints"
      },
      "bearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication (Bearer djd_live_...)"
      }
    }
  }
}